refactor: update import paths to use @server/models
