+++ apps/server/src/models/auth.ts
