+++ apps/server/src/models/errors.ts
