+++ apps/server/src/models/generalPreference.ts
@@ -1,5 +1,7 @@
 import { z } from 'zod';
 
+import { $Enums } from '@prisma/client';
+
 export enum Appearance {
   LIGHT = 'light',
   DARK = 'dark',
@@ -31,11 +33,11 @@ export enum Layout {
 export const generalPreferencesSchema = z.object({
   id: z.number().int().nonnegative(),
   userId: z.number().int().nonnegative(),
-  appearance: z.nativeEnum(Appearance),
-  density: z.nativeEnum(Density),
-  openFiles: z.nativeEnum(OpenFiles),
-  layout: z.nativeEnum(Layout),
-  language: z.nativeEnum(Language)
+  appearance: z.nativeEnum($Enums.appearanceEnum),
+  density: z.nativeEnum($Enums.densityEnum),
+  openFiles: z.nativeEnum($Enums.openFilesEnum),
+  layout: z.nativeEnum($Enums.layoutEnum),
+  language: z.nativeEnum($Enums.languageEnum)
 });
 
 export type GeneralPreferences = z.infer<typeof generalPreferencesSchema>;
